chore: reorg file structure and add jsdoc
