+++ src/types/aiApiClient/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export * from './types/PromptStoreEntry';
+export * from './types/TelemetryData';
+export * from './types/ApiClient';
+export * from './enums/commandSource';
+export * from './utils/Completion/Completion';
+export * from './Completions/AiCompletion';
+export * from './Completions/InlineCompletions';
